requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
numpy>=1.24
streamlit>=1.28.0
groq>=0.4.0
python-dotenv>=1.0.0
